Removing strsrch to make text reading more stable
Let me try to explain why strsrch is bad. You have 'Kosinski', and then
you have 'Moduled Kosinski'. Now, because they both contain 'Kosinski',
the latter can be mistaken for the former, depending on if 'Kosinski'
comes earlier in comprTypes.

The reason is because strsrch doesn't really have an explicit starting
point: it scans for a string starting at the first character, and slowly
moving on from there. That means, to it, 'febfeshfioseKosinskifafwaf' ==
'Kosinski'. Yes, that is good if the user is liberal with their spacing
('    Kosinski  '), but that's what trimString is for. Anyway, this
causes issues with Moduled Kosinski support, so I'm using something more
strict.