Improving BGR->RGB conversion (and fixing an old bug)
For some reason, SDL2 doesn't accept non-standard colour formats, which is why my SDL2 port took so long to be made, since the original SDL1 code relied on it. This new version converts the palette entries as the file's loaded, rather than during tile generation.

This also fixes a bug that's been around since the original PlaneEd: DrawTileBlank reads from the internal palette as if it's RGB, but, until now, it was BGR. This means 'blank' tiles like the ones between the 'SONIC's and 'MILES's in Sonic 2's menu background now use the correct colour.